# lib/api-zod/src/generated/types/playUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * The game show's API. Public routes serve the main screen and sign-in; every /admin route needs the session cookie set by /auth/setup or /auth/login.
 * OpenAPI spec version: 1.0.0
 */

/**
 * Omit teamId to record that nobody won this tile.
 */
export interface PlayUpdate {
  teamId?: number;
  /** @minimum 0 */
  points: number;
}